+++ src/invoices/VoucherExport/BatchVoucherExportsList/index.js
@@ -1,0 +1,1 @@
+export { BatchVoucherExportsList } from './BatchVoucherExportsList';