+++ src/invoices/VoucherExport/utils/index.js
@@ -1,0 +1,1 @@
+export { createManualVoucherExport } from './createManualVoucherExport';