+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * On small screens the action menu is covered by the header. Refs UINV-374.
 * Add "Voucher export" to actions menu. Refs UINV-368.
 * Add "Export results (CSV)" action to invoice app. Refs UINV-383.
+* Select batch group and display existing exports. Refs UINV-369.
+* Run manual export and display toast. Refs UINV-370.
+* Download batch export files from full screen view. Refs UINV-371.
 
 ## [3.1.1](https://github.com/folio-org/ui-invoice/tree/v3.1.1) (2022-03-22)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.1.0...v3.1.1)